+++ examples/api.ts
@@ -15,19 +15,22 @@ export interface ComplexData {
   levels: Array<Array<FoodItem>>;
 }
 
-export const getList = async (
-  itemCount?: number,
-): Promise<Response<Array<FoodItem>>> => {
-  itemCount =
-    itemCount !== undefined ? itemCount : 2 + Math.ceil(Math.random() * 5);
+export const getFoodList = async (options?: {
+  itemCount?: number;
+  search?: string;
+}): Promise<Response<Array<FoodItem>>> => {
+  const itemCount = options?.itemCount !== undefined ? options.itemCount : 2 + Math.ceil(Math.random() * 5);
   const set = new Set<number>();
-  while (set.size < itemCount && set.size !== food.length) {
-    set.add(Math.floor(Math.random() * food.length));
+  const foodList = food.filter(
+    (item) => !options?.search || item.name.toLowerCase().includes(options.search.toLowerCase()),
+  );
+  while (set.size < itemCount && set.size !== foodList.length) {
+    set.add(Math.floor(Math.random() * foodList.length));
   }
 
   return {
     status: 200,
-    data  : Array.from(set).map(index => food[index]),
+    data: Array.from(set).map((index) => foodList[index]),
   };
 };
 
@@ -37,16 +40,16 @@ export const getComplexData = async (): Promise<Response<ComplexData>> => {
       .repeat(10)
       .split("")
       .map(async () => {
-        const response = await getList(2 + Math.floor(Math.random() * 10));
+        const response = await getFoodList({ itemCount: 2 + Math.floor(Math.random() * 10) });
         return response.data;
       }),
   );
 
   return {
     status: 200,
-    data  : {
+    data: {
       currentLevel: 3,
-      levels      : levels,
+      levels: levels,
     },
   };
 };