+++ src/Updater.ts
@@ -1,7 +1,7 @@
-import {Dispatch, SetStateAction} from "react";
-import AsyncState                 from "./AsyncState";
-import {PromiseOrAsyncFunction}   from "./Types";
-import AsyncStateGroup            from "./AsyncStateGroup";
+import type { Dispatch, SetStateAction } from "react";
+import AsyncState from "./AsyncState";
+import AsyncStateGroup from "./AsyncStateGroup";
+import type { PromiseOrAsyncFunction } from "./Types";
 
 export interface UpdateAsyncStateOptions {
   /** do not reset value when submitting */
@@ -11,15 +11,17 @@ export interface UpdateAsyncStateOptions {
   /** if update takes longer than `timeout` ms: reject */
   timeout?: number;
 
-  errorDebug?: boolean,
+  errorDebug?: boolean;
 }
 
-export const updateAsyncState = async <T, A extends AsyncState<T, any>>(setAsyncState: Dispatch<SetStateAction<A>>,
-                                                                        promiseOrAsyncFn: PromiseOrAsyncFunction<T>,
-                                                                        options?: UpdateAsyncStateOptions): Promise<A> => {
+export const updateAsyncState = async <T, A extends AsyncState<T, any>>(
+  setAsyncState: Dispatch<SetStateAction<A>>,
+  promiseOrAsyncFn: PromiseOrAsyncFunction<T>,
+  options?: UpdateAsyncStateOptions,
+): Promise<A> => {
   let pendingAt: number | null = null;
 
-  setAsyncState(currentState => {
+  setAsyncState((currentState) => {
     const newState = options?.refresh ? AsyncState.refresh(currentState) : AsyncState.submit(currentState);
     pendingAt = newState.pendingAt;
     return newState as A;
@@ -29,10 +31,9 @@ export const updateAsyncState = async <T, A extends AsyncState<T, any>>(setAsync
   let resolved = false;
   let possiblyResolve: (stateUpdater: (state: AsyncState<T>) => AsyncState<T>) => boolean = () => {
     throw new Error("[react-async-stateful] This should never happen!");
-    return false;
   };
 
-  const valuePromise = new Promise<A>(resolve => {
+  const valuePromise = new Promise<A>((resolve) => {
     possiblyResolve = (stateUpdater) => {
       if (resolved) {
         return false;
@@ -42,7 +43,7 @@ export const updateAsyncState = async <T, A extends AsyncState<T, any>>(setAsync
         clearTimeout(timeoutId);
       }
       resolved = true;
-      setAsyncState(currentState => {
+      setAsyncState((currentState) => {
         if (currentState.pendingAt !== pendingAt) {
           // this state has been updated again since we started -- we are no longer in charge of resolving this state and shouldn't update it
           return currentState;
@@ -59,23 +60,22 @@ export const updateAsyncState = async <T, A extends AsyncState<T, any>>(setAsync
   if (options?.timeout) {
     timeoutId = setTimeout(() => {
       const error = new Error(`[react-async-stateful] Async update timed out, took longer than ${options.timeout}ms`);
-      if (possiblyResolve(state => AsyncState.reject(state, error))) {
+      if (possiblyResolve((state) => AsyncState.reject(state, error))) {
         console.error(error);
       }
     }, options.timeout);
   }
 
   try {
-    const promise =
-            typeof promiseOrAsyncFn === "function"
-              ? promiseOrAsyncFn()
-              : promiseOrAsyncFn;
+    const promise = typeof promiseOrAsyncFn === "function" ? promiseOrAsyncFn() : promiseOrAsyncFn;
     if (typeof promise.then !== "function") {
-      throw new Error(`[react-async-stateful] ${
-        typeof promiseOrAsyncFn === "function"
-          ? "Function provided did not return a promise"
-          : "First argument was not a promise or an async function"
-      }`);
+      throw new Error(
+        `[react-async-stateful] ${
+          typeof promiseOrAsyncFn === "function"
+            ? "Function provided did not return a promise"
+            : "First argument was not a promise or an async function"
+        }`,
+      );
     }
 
     let minimumPendingPromise: Promise<void> | undefined = undefined;
@@ -89,9 +89,9 @@ export const updateAsyncState = async <T, A extends AsyncState<T, any>>(setAsync
       value = null as any;
     }
 
-    possiblyResolve(state => AsyncState.resolve(state, value));
-  } catch (error) {
-    if (possiblyResolve(state => AsyncState.reject(state, error))) {
+    possiblyResolve((state) => AsyncState.resolve(state, value));
+  } catch (error: any) {
+    if (possiblyResolve((state) => AsyncState.reject(state, error))) {
       console.error(`[react-async-stateful] Updating async state failed: ${error.stack || error.message}`);
 
       if (options?.errorDebug) {
@@ -109,28 +109,30 @@ export const updateAsyncStateElement = async <T extends K, A extends AsyncStateG
   setAsyncState: Dispatch<SetStateAction<A>>,
   key: K,
   promiseOrAsyncFn: PromiseOrAsyncFunction<T>,
-  options?: UpdateAsyncStateOptions): Promise<A> => {
+  options?: UpdateAsyncStateOptions,
+): Promise<A> => {
   let valueResolve: (state: A) => void = () => {
     throw new Error("This should never happen!");
   };
-  const valuePromise = new Promise<A>(resolve => {
+  const valuePromise = new Promise<A>((resolve) => {
     valueResolve = resolve;
   });
 
-  const promise =
-          typeof promiseOrAsyncFn === "function"
-            ? promiseOrAsyncFn()
-            : promiseOrAsyncFn;
+  const promise = typeof promiseOrAsyncFn === "function" ? promiseOrAsyncFn() : promiseOrAsyncFn;
 
   const update = async () => {
-    const updateElement = (elementOrUpdater: AsyncState<T> | ((prevElement: AsyncState<T>) => AsyncState<T>)) => setAsyncState((prevState => {
-      const element = typeof elementOrUpdater === "function" ? elementOrUpdater(AsyncStateGroup.getOrCreateElement(prevState, key)) : elementOrUpdater;
-      const updatedState = AsyncStateGroup.setElement(prevState, key, element) as A;
-      if (updatedState.resolved || updatedState.rejected) {
-        valueResolve(updatedState);
-      }
-      return updatedState;
-    }));
+    const updateElement = (elementOrUpdater: AsyncState<T> | ((prevElement: AsyncState<T>) => AsyncState<T>)) =>
+      setAsyncState((prevState) => {
+        const element =
+          typeof elementOrUpdater === "function"
+            ? elementOrUpdater(AsyncStateGroup.getOrCreateElement(prevState, key))
+            : elementOrUpdater;
+        const updatedState = AsyncStateGroup.setElement(prevState, key, element) as A;
+        if (updatedState.resolved || updatedState.rejected) {
+          valueResolve(updatedState);
+        }
+        return updatedState;
+      });
     await updateAsyncState(updateElement, promise, options);
   };