+++ src/Utils.ts
@@ -4,25 +4,26 @@ export type NotFunction = any & {
 };
 
 // polyfill from https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Object/assign
-export const assign: typeof Object["assign"] = Object.assign || function(target: any, varArgs: any) {
-  if (target === null || target === undefined) {
-    throw new TypeError("Cannot convert undefined or null to object");
-  }
+export const assign: (typeof Object)["assign"] =
+  Object.assign ||
+  ((target: any, ...varArgs: any) => {
+    if (target === null || target === undefined) {
+      throw new TypeError("Cannot convert undefined or null to object");
+    }
 
-  const to = Object(target);
+    const to = Object(target);
 
-  for (let index = 1; index < arguments.length; index++) {
-    // eslint-disable-next-line prefer-rest-params
-    const nextSource = arguments[index];
+    for (let index = 1; index < varArgs.length; index++) {
+      const nextSource = varArgs[index];
 
-    if (nextSource !== null && nextSource !== undefined) {
-      for (const nextKey in nextSource) {
-        // Avoid bugs when hasOwnProperty is shadowed
-        if (Object.prototype.hasOwnProperty.call(nextSource, nextKey)) {
-          to[nextKey] = nextSource[nextKey];
+      if (nextSource !== null && nextSource !== undefined) {
+        for (const nextKey in nextSource) {
+          // Avoid bugs when hasOwnProperty is shadowed
+          if (Object.prototype.hasOwnProperty.call(nextSource, nextKey)) {
+            to[nextKey] = nextSource[nextKey];
+          }
         }
       }
     }
-  }
-  return to;
-};
+    return to;
+  });