+++ src/redux/index.ts
@@ -1,2 +1,0 @@
-export * from "./Redux";
-export * from "./ReduxTypes";